+++ core/execution/engine.go
@@ -114,6 +114,8 @@ type Engine struct {
 	successorWindow time.Duration
 	// only used once, during CP restore, this doesn't need to be included in a snapshot or checkpoint.
 	skipRestoreSuccessors map[string]struct{}
+
+	delayTransactionsTarget common.DelayTransactionsTarget
 }
 
 // NewEngine takes stores and engines and returns
@@ -132,6 +134,7 @@ func NewEngine(
 	volumeDiscountService fee.VolumeDiscountService,
 	banking common.Banking,
 	parties common.Parties,
+	delayTransactionsTarget common.DelayTransactionsTarget,
 ) *Engine {
 	// setup logger
 	log = log.Named(namedLogger)
@@ -159,6 +162,7 @@ func NewEngine(
 		volumeDiscountService:         volumeDiscountService,
 		banking:                       banking,
 		parties:                       parties,
+		delayTransactionsTarget:       delayTransactionsTarget,
 	}
 
 	// set the eligibility for proposer bonus checker
@@ -561,6 +565,7 @@ func (e *Engine) UpdateSpotMarket(ctx context.Context, marketConfig *types.Marke
 	if err := mkt.Update(ctx, marketConfig); err != nil {
 		return err
 	}
+	e.delayTransactionsTarget.MarketDelayRequiredUpdated(mkt.GetID(), marketConfig.EnableTxReordering)
 	e.publishUpdateMarketInfos(ctx, mkt.GetMarketData(), *mkt.Mkt())
 	return nil
 }
@@ -618,6 +623,7 @@ func (e *Engine) UpdateMarket(ctx context.Context, marketConfig *types.Market) e
 	if err := mkt.Update(ctx, marketConfig, e.oracle); err != nil {
 		return err
 	}
+	e.delayTransactionsTarget.MarketDelayRequiredUpdated(mkt.GetID(), marketConfig.EnableTxReordering)
 	e.publishUpdateMarketInfos(ctx, mkt.GetMarketData(), *mkt.Mkt())
 	return nil
 }
@@ -700,7 +706,7 @@ func (e *Engine) submitMarket(ctx context.Context, marketConfig *types.Market, o
 		)
 		return err
 	}
-
+	e.delayTransactionsTarget.MarketDelayRequiredUpdated(mkt.GetID(), marketConfig.EnableTxReordering)
 	e.futureMarkets[marketConfig.ID] = mkt
 	e.futureMarketsCpy = append(e.futureMarketsCpy, mkt)
 	e.allMarkets[marketConfig.ID] = mkt
@@ -780,7 +786,7 @@ func (e *Engine) submitSpotMarket(ctx context.Context, marketConfig *types.Marke
 		)
 		return err
 	}
-
+	e.delayTransactionsTarget.MarketDelayRequiredUpdated(mkt.GetID(), marketConfig.EnableTxReordering)
 	e.spotMarkets[marketConfig.ID] = mkt
 	e.spotMarketsCpy = append(e.spotMarketsCpy, mkt)
 	e.allMarkets[marketConfig.ID] = mkt