+++ core/execution/snapshot_test.go
@@ -630,7 +630,8 @@ func getEngine(t *testing.T, vegaPath paths.Paths, now time.Time) *snapshotTestD
 	referralDiscountReward.EXPECT().GetReferrer(gomock.Any()).Return(types.PartyID(""), errors.New("not a referrer")).AnyTimes()
 	banking := mocks.NewMockBanking(ctrl)
 	parties := mocks.NewMockParties(ctrl)
-
+	delayTarget := mocks.NewMockDelayTransactionsTarget(ctrl)
+	delayTarget.EXPECT().MarketDelayRequiredUpdated(gomock.Any(), gomock.Any()).AnyTimes()
 	eng := execution.NewEngine(
 		log,
 		cfg,
@@ -645,6 +646,7 @@ func getEngine(t *testing.T, vegaPath paths.Paths, now time.Time) *snapshotTestD
 		volumeDiscount,
 		banking,
 		parties,
+		delayTarget,
 	)
 
 	statsData := stats.New(log, stats.NewDefaultConfig())
@@ -701,7 +703,8 @@ func getEngineWithParties(t *testing.T, now time.Time, balance *num.Uint, partie
 	referralDiscountReward.EXPECT().GetReferrer(gomock.Any()).Return(types.PartyID(""), errors.New("not a referrer")).AnyTimes()
 	banking := mocks.NewMockBanking(ctrl)
 	partiesMock := mocks.NewMockParties(ctrl)
-
+	delayTarget := mocks.NewMockDelayTransactionsTarget(ctrl)
+	delayTarget.EXPECT().MarketDelayRequiredUpdated(gomock.Any(), gomock.Any()).AnyTimes()
 	eng := execution.NewEngine(
 		log,
 		cfg,
@@ -716,6 +719,7 @@ func getEngineWithParties(t *testing.T, now time.Time, balance *num.Uint, partie
 		volumeDiscount,
 		banking,
 		partiesMock,
+		delayTarget,
 	)
 
 	statsData := stats.New(log, stats.NewDefaultConfig())